+++ owrx/source/airspy.py
@@ -5,7 +5,12 @@ from .soapy import SoapyConnectorSource
 class AirspySource(SoapyConnectorSource):
     def getSoapySettingsMappings(self):
         mappings = super().getSoapySettingsMappings()
-        mappings.update({"bias_tee": "biastee"})
+        mappings.update(
+            {
+                "bias_tee": "biastee",
+                "bitpack": "bitpack",
+            }
+        )
         return mappings
 
     def getDriver(self):